Move zoom buttons in story editor to top toolbar
